+++ src/main/java/org/doancnpm/Models/PhieuXuat.java
@@ -12,7 +12,7 @@ public class PhieuXuat {
     Integer maDaiLy;
     Integer maNhanVien;
     Date ngayLapPhieu;
-    Double tongTien;
+    Long tongTien;
     String ghiChu;
 
     public Integer getID() {
@@ -55,11 +55,11 @@ public class PhieuXuat {
         this.ngayLapPhieu = ngayLapPhieu;
     }
 
-    public Double getTongTien() {
+    public Long getTongTien() {
         return tongTien;
     }
 
-    public void setTongTien(Double tongTien) {
+    public void setTongTien(Long tongTien) {
         this.tongTien = tongTien;
     }
